created component Counter and View Information
