+++ pages/_app.js
@@ -1,5 +1,4 @@
 import '../styles/globals.css'
-import 'tailwindcss/tailwind.css'
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />
